setup elements
