+++ package.json
@@ -13,6 +13,8 @@
     "@heroicons/react": "^2.1.1",
     "@nextui-org/react": "^2.2.9",
     "@reduxjs/toolkit": "^2.2.1",
+    "@stripe/react-stripe-js": "^2.6.0",
+    "@stripe/stripe-js": "^3.0.8",
     "framer-motion": "^11.0.7",
     "leaflet": "^1.9.4",
     "react": "^18.2.0",
