+++ src/main.tsx
@@ -7,7 +7,8 @@ import Home from "./Pages/Home";
 import Shop from "./Pages/Shop.jsx";
 import ProductPage from "./Pages/ProductPage.jsx";
 import Cart from "./Pages/Cart.js";
-import Checkout from "./Pages/Checkout.jsx";
+import Order from "./Pages/Order.js";
+import Checkout from "./Pages/Checkout.tsx";
 import ThankYou from "./Pages/ThankYou.jsx";
 import Account from "./Pages/Account.jsx";
 
@@ -29,7 +30,8 @@ const router = createBrowserRouter(
       <Route path="/shop" element={<Shop />} />
       <Route path="/shop/:slug" element={<ProductPage />} />
       <Route path="/cart" element={<Cart />} />
-      <Route path="/cart/pass-your-order" element={<Checkout />} />
+      <Route path="/cart/pass-your-order" element={<Order />} />
+      <Route path="/cart/checkout" element={<Checkout />} />
       <Route path="/cart/pass-your-order/success" element={<ThankYou />} />
       <Route path="/account" element={<Account />} />
     </Route>
